py/requirements.txt
pyserial
libscrc